Refactor utils folder
